split up scripts.js into separate modules in utils/
