--- frontend/www/scripts/defaults.js
@@ -1,1 +1,0 @@
-const HOST = "http://localhost:8000";
